+++ hdsr_fewspy/api_calls/time_series/base.py
@@ -3,8 +3,10 @@ from datetime import datetime
 from hdsr_fewspy import exceptions
 from hdsr_fewspy.api_calls.base import GetRequest
 from hdsr_fewspy.constants.choices import ApiParameters
+from hdsr_fewspy.constants.choices import DefaultPiSettingsChoices
 from hdsr_fewspy.constants.choices import PiRestDocumentFormatChoices
 from hdsr_fewspy.constants.custom_types import ResponseType
+from hdsr_fewspy.constants.pi_settings import PiSettings
 from hdsr_fewspy.converters.utils import datetime_to_fews_date_str
 from hdsr_fewspy.converters.utils import fews_date_str_to_datetime
 from hdsr_fewspy.converters.xml_to_python_obj import parse
@@ -27,6 +29,9 @@ class GetTimeSeriesBase(GetRequest):
     response_text_location_not_found = "Some of the location ids do not exist"
     response_text_parameter_not_found = "Some of the parameters do not exists"
 
+    start_time_all = datetime_to_fews_date_str(date_time=datetime(year=2011, month=1, day=1))
+    end_time_all = datetime_to_fews_date_str(date_time=datetime.now())
+
     def __init__(
         self,
         start_time: Union[datetime, str],
@@ -46,7 +51,13 @@ class GetTimeSeriesBase(GetRequest):
         self.end_time: datetime = self.__validate_time(time=end_time)
         self.location_ids = location_ids
         self.parameter_ids = parameter_ids
-        self.qualifier_ids = qualifier_ids
+
+        # TODO: refactor this organic ugly code
+        retry_backoff_session = kwargs.get("retry_backoff_session")
+
+        self.qualifier_ids = self.__validate_qualifier(
+            qualifier_ids=qualifier_ids, pi_settings=retry_backoff_session.pi_settings
+        )
         self.thinning = thinning
         self.omit_empty_time_series = omit_empty_time_series
         self.drop_missing_values = drop_missing_values
@@ -62,6 +73,31 @@ class GetTimeSeriesBase(GetRequest):
     def __validate_time(time: Union[datetime, str]) -> datetime:
         datetime_obj = fews_date_str_to_datetime(fews_date_str=time) if isinstance(time, str) else time
         return datetime_obj
+
+    @staticmethod
+    def __validate_qualifier(qualifier_ids: List[str], pi_settings: PiSettings) -> List[str]:
+        """For all area related get_time_series we require a qualifier to avoid a response with >1 time-series."""
+        mapper = {
+            DefaultPiSettingsChoices.wis_production_area_soilmoisture.value: ["Lband05cm", "Lband10cm", "Lband20cm"],
+            DefaultPiSettingsChoices.wis_production_area_precipitation_wiwb.value: ["wiwb_merge"],
+            DefaultPiSettingsChoices.wis_production_area_precipitation_radarcorrection.value: ["mfbs_merge"],
+            #
+            DefaultPiSettingsChoices.wis_production_area_evaporation_wiwb_satdata.value: ["RA", "satdata_merge", ""],
+            DefaultPiSettingsChoices.wis_production_area_evaporation_waterwatch.value: [""],
+        }
+        required_qualifiers = mapper.get(pi_settings.settings_name, None)  # noqa
+        if not required_qualifiers:
+            return qualifier_ids
+        assert qualifier_ids is not None, (
+            f"pi_settings '{pi_settings.settings_name}' get_time_series can only be used with a "
+            f"qualifier_id. Choose from {required_qualifiers}"
+        )
+        qualifier_ids_list = [qualifier_ids] if not isinstance(qualifier_ids, list) else qualifier_ids
+        for qualifier_id in qualifier_ids_list:
+            assert (
+                qualifier_id in required_qualifiers
+            ), f"qualifier_id '{qualifier_id}' must be in {required_qualifiers}"
+        return qualifier_ids
 
     @property
     def url_post_fix(self):
@@ -100,6 +136,14 @@ class GetTimeSeriesBase(GetRequest):
             ApiParameters.start_time,
         ]
 
+    @staticmethod
+    def get_task_uuid(request_params: Dict) -> str:
+        loc = request_params.get("locationIds", "")
+        par = request_params.get("parameterIds", "")
+        qual = request_params.get("qualifierIds", "")
+        task_uuid = f"{loc} {par} {qual}"
+        return task_uuid.strip()
+
     def _download_time_series(
         self,
         date_ranges: List[Tuple[pd.Timestamp, pd.Timestamp]],
@@ -114,6 +158,19 @@ class GetTimeSeriesBase(GetRequest):
         (smaller or larger windows) parameters 'startTime' and 'endTime' again.
         """
         responses = responses if responses else []
+
+        # firstly, check if any time-series exist at all (no start_time and end_time)
+        try:
+            has_time_series = self._get_nr_timestamps_no_start_end(request_params=request_params) > 0
+            if not has_time_series:
+                task_uuid = self.get_task_uuid(request_params=request_params)
+                logger.info(f"skipping since no time-series at all for '{task_uuid}'")
+                return []
+        except (exceptions.LocationIdsDoesNotExistErr, exceptions.ParameterIdsDoesNotExistErr) as err:
+            logger.warning(err)
+            return []
+
+        # secondly, download time-series in little chunks
         for request_index, (data_range_start, data_range_end) in enumerate(date_ranges):
             # update start and end in request params
             request_params["startTime"] = datetime_to_fews_date_str(data_range_start)
@@ -148,12 +205,7 @@ class GetTimeSeriesBase(GetRequest):
                     responses=responses,
                 )
             else:
-                DateFrequencyBuilder.log_progress_download_ts(
-                    data_range_start=data_range_start,
-                    data_range_end=data_range_end,
-                    ts_end=pd.Timestamp(self.end_time),
-                )
-                # ready to download timeseries (with new_date_range_freq)
+                # ready to download time-series (with new_date_range_freq)
                 request_params["onlyHeaders"] = False
                 request_params["showStatistics"] = False
                 response = self.retry_backoff_session.get(
@@ -163,6 +215,12 @@ class GetTimeSeriesBase(GetRequest):
                     logger.error(f"FEWS Server responds {response.text}")
                 else:
                     responses.append(response)
+                DateFrequencyBuilder.log_progress_download_ts(
+                    task=self.get_task_uuid(request_params=request_params),
+                    request_end=data_range_end,
+                    ts_start=pd.Timestamp(self.start_time),
+                    ts_end=pd.Timestamp(self.end_time),
+                )
         return responses
 
     def _get_statistics(self, request_params: Dict) -> ResponseType:
@@ -173,8 +231,14 @@ class GetTimeSeriesBase(GetRequest):
         )
         return response
 
+    def _get_nr_timestamps_no_start_end(self, request_params: Dict) -> int:
+        request_params_copy = request_params.copy()
+        request_params_copy["startTime"] = self.start_time_all
+        request_params_copy["endTime"] = self.end_time_all
+        return self._get_nr_timestamps(request_params_copy)
+
     def _get_nr_timestamps(self, request_params: Dict) -> int:
-        assert "moduleInstanceIds" in request_params, "code error GetTimeSeriesBase"
+        assert "moduleInstanceIds" in request_params, "code error _get_nr_timestamps"
         response = self._get_statistics(request_params=request_params)
         msg = f"status_code={response.status_code}, err={response.text}, request_params={request_params}"
         if not response.ok: