+++ hdsr_fewspy/constants/choices.py
@@ -1,31 +1,76 @@
+from __future__ import annotations
+
+from enum import Enum
 from typing import List
 from typing import Union
 
 
-class PiRestDocumentFormatChoices:
+class PiRestDocumentFormatChoices(Enum):
     # TODO: make this a enum.Enum class to improve type hints trough codebase
     xml = "PI_XML"
     json = "PI_JSON"
 
     @classmethod
-    def get_all(cls) -> List[str]:
-        return [cls.xml, cls.json]
+    def get_all(cls):
+        return [x for x in cls.__members__.values()]
 
 
-class DefaultPiSettingsChoices:
-    # TODO: make this a enum.Enum class to improve type hints trough codebase
-    wis_production = "wis_production"
-    wis_stand_alone = "wis_stand_alone"
-    # TODO: efcis_production = "efcis_production"
-    # TODO: efcis_stand_alone = "efcis_stand_alone"
+class DefaultPiSettingsChoices(Enum):
+    """
+    Choices that must be present in https://github.com/hdsr-mid/hdsr_fewspy_auth/blob/main/settings.csv.
+
+    Each choice has a different module_instance_id. We specify module_instance_id in settings.csv, to avoid
+    requesting 3 time-series for one location. The module_instance_id's are:
+     - ImportOpvlWater = raw data
+     - WerkFilter = this data is being validated by HDSR person (data validator CAW). This data might change every day
+     - MetingenFilter = validated data  reeksen. This data is months behind the current situation.
+    """
+
+    # TODO: add efcis choices
+    wis_production_point_raw = "wis_production_point_raw"
+    wis_production_point_work = "wis_production_point_work"
+    wis_production_point_validated = "wis_production_point_validated"
+    #
+    wis_production_area_soilmoisture = "wis_production_area_soilmoisture"
+    # module_instance_id=ImportBodemvocht,
+    # parameter= e.g. SM.d
+    # qualifiers=Lband05cm, Lband10cm, Lband20cm
+    # noqa filterId=INTERNAL-API.GEBGEM&locationIds=AFVG13&parameterIds=SM.d&moduleInstanceIds=ImportBodemvocht&startTime=2020-01-01T00%3A00%3A00Z&endTime=2023-01-01T00%3A00%3A00Z&onlyHeaders=true&showStatistics=true
+    #
+    wis_production_area_precipitation_wiwb = "wis_production_area_precipitation_wiwb"
+    # module_instance_id ImportGridHistWiwb tm 2019
+    # parameter = e.g. Rh.h
+    # qualifiers=wiwb_merge
+    # noqa filterId=INTERNAL-API.GEBGEM&locationIds=AFVG13&parameterIds=Rh.h&moduleInstanceIds=ImportGridHistWiwb&startTime=2015-01-01T00%3A00%3A00Z&endTime=2023-01-01T00%3A00%3A00Z&onlyHeaders=true&showStatistics=true
+    #
+    wis_production_area_precipitation_radarcorrection = "wis_production_area_precipitation_radarcorrection"
+    # module_instance_id RadarCorrectie vanaf 2019
+    # parameter = e.g. Rh.h
+    # qualifiers=mfbs_merge
+    # noqa filterId=INTERNAL-API.GEBGEM&locationIds=AFVG13&parameterIds=Rh.h&moduleInstanceIds=RadarCorrectie&startTime=2015-01-01T00%3A00%3A00Z&endTime=2023-01-01T00%3A00%3A00Z&onlyHeaders=true&showStatistics=true
+
+    wis_production_area_evaporation_wiwb_satdata = "wis_production_area_evaporation_wiwb_satdata"
+    # module_instance_id=ImportWiwbSatData
+    # parameter= e.g. Eact.d
+    # qualifiers=satdata_merge, RA, None
+    # noqa filterId=INTERNAL-API.GEBGEM&locationIds=AFVG13&parameterIds=Eact.d&moduleInstanceIds=ImportWiwbSatData&startTime=2015-01-01T00%3A00%3A00Z&endTime=2023-01-01T00%3A00%3A00Z&onlyHeaders=true&showStatistics=true
+
+    wis_production_area_evaporation_waterwatch = "wis_production_area_evaporation_waterwatch"
+    # module_instance_id=ImportWaterwatch
+    # parameter= e.g. Eact.d
+    # qualifiers=None
+    # noqa filterId=INTERNAL-API.GEBGEM&locationIds=AFVG13&parameterIds=Eact.d&moduleInstanceIds=ImportWaterwatch&startTime=2020-01-01T00%3A00%3A00Z&endTime=2023-01-01T00%3A00%3A00Z
+
+    wis_stand_alone_point_raw = "wis_stand_alone_point_raw"
+    wis_stand_alone_point_work = "wis_stand_alone_point_work"
+    wis_stand_alone_point_validated = "wis_stand_alone_point_validated"
 
     @classmethod
-    def get_all(cls) -> List[str]:
-        return [cls.wis_production, cls.wis_stand_alone]
+    def get_all(cls) -> List[DefaultPiSettingsChoices]:
+        return [x for x in cls.__members__.values()]
 
 
-class OutputChoices:
-    # TODO: make this a enum.Enum class to improve type hints trough codebase
+class OutputChoices(Enum):
     xml_file_in_download_dir = "xml_file_in_download_dir"
     json_file_in_download_dir = "json_file_in_download_dir"
     csv_file_in_download_dir = "csv_file_in_download_dir"
@@ -34,19 +79,19 @@ class OutputChoices:
     pandas_dataframe_in_memory = "pandas_dataframe_in_memory"
 
     @classmethod
-    def validate(cls, output_choice: str) -> str:
-        assert output_choice in cls.get_all(), f"output_choice {output_choice} must be in {cls.get_all()}"
+    def validate(cls, output_choice: OutputChoices) -> OutputChoices:
+        assert isinstance(output_choice, OutputChoices)
         return output_choice
 
     @classmethod
-    def get_pi_rest_document_format(cls, output_choice: str) -> str:
+    def get_pi_rest_document_format(cls, output_choice: OutputChoices) -> PiRestDocumentFormatChoices:
         output_choice = cls.validate(output_choice=output_choice)
         if output_choice in [cls.xml_file_in_download_dir, cls.xml_response_in_memory]:
             return PiRestDocumentFormatChoices.xml
         return PiRestDocumentFormatChoices.json
 
     @classmethod
-    def needs_output_dir(cls, output_choice: str) -> bool:
+    def needs_output_dir(cls, output_choice: OutputChoices) -> bool:
         cls.validate(output_choice=output_choice)
         return output_choice in {
             cls.xml_file_in_download_dir,
@@ -55,29 +100,25 @@ class OutputChoices:
         }
 
     @classmethod
-    def get_all(cls) -> List[str]:
-        return [
-            cls.xml_file_in_download_dir,
-            cls.json_file_in_download_dir,
-            cls.csv_file_in_download_dir,
-            cls.xml_response_in_memory,
-            cls.json_response_in_memory,
-            cls.pandas_dataframe_in_memory,
-        ]
+    def get_all(cls) -> List[OutputChoices]:
+        return [x for x in cls.__members__.values()]
 
 
-class TimeZoneChoices:
-    # TODO: make this a enum.Enum class to improve type hints trough codebase
+class TimeZoneChoices(Enum):
     gmt = 0.0  # ["GMT"= "Etc/GMT" = "Etc/GMT-0"]
     eu_amsterdam = 1.0  # ["Europe/Amsterdam"]
 
     @classmethod
-    def get_hdsr_default(cls) -> float:
+    def get_hdsr_default(cls):
         return cls.gmt
 
     @classmethod
-    def get_all(cls) -> List[float]:
-        return [cls.gmt, cls.eu_amsterdam]
+    def get_all(cls) -> List[TimeZoneChoices]:
+        return [x for x in cls.__members__.values()]
+
+    @classmethod
+    def get_all_values(cls) -> List[float]:
+        return [x.value for x in cls.__members__.values()]
 
     @classmethod
     def date_string_format(cls) -> str:
@@ -86,11 +127,11 @@ class TimeZoneChoices:
     @classmethod
     def get_tz_float(cls, value: Union[str, float, int]) -> float:
         mapper = {
-            "GMT": cls.gmt,
-            "GMT-0": cls.gmt,
-            "Etc/GMT": cls.gmt,
-            "Etc/GMT-0": cls.gmt,
-            "Europe/Amsterdam": cls.eu_amsterdam,
+            "GMT": cls.gmt.value,
+            "GMT-0": cls.gmt.value,
+            "Etc/GMT": cls.gmt.value,
+            "Etc/GMT-0": cls.gmt.value,
+            "Europe/Amsterdam": cls.eu_amsterdam.value,
         }
         if isinstance(value, str):
             if value in mapper.keys():
@@ -99,8 +140,8 @@ class TimeZoneChoices:
         try:
             float_value = float(value)
         except ValueError:
-            raise AssertionError(f"{msg} value {value}")
-        assert float_value in cls.get_all(), f"{msg} value {value} and float_value {float_value}"
+            raise AssertionError(f"{msg} value '{value}'")
+        assert float_value in cls.get_all_values(), f"{msg} value '{value}' and float_value '{float_value}'"
         return float_value